+++ app/src/main/java/de/uni_bremen/comnets/resourcemonitor/ServerCommunicationHandler.java
@@ -19,9 +19,6 @@ import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
 
-import javax.net.ssl.HttpsURLConnection;
-
-
 /**
  * Class to handle the data upload to the server
  */